+++ message/sync_message.go
@@ -97,6 +97,7 @@ func (m SyncInserts) GetRecords() []arrow.Record {
 	return res
 }
 
+// Get all records for a single table
 func (m SyncInserts) GetRecordsForTable(table *schema.Table) []arrow.Record {
 	res := make([]arrow.Record, 0, len(m))
 	for _, insert := range m {