+++ plugin/sort.go
@@ -11,6 +11,7 @@ import (
 // sort records by specific column. This is intended for testing purposes only.
 // Because "id" is auto-incrementing in the test  data generator, if passed "id"
 // this should result in records being returned in insertion order.
+// nolint:unparam
 func sortRecords(table *schema.Table, records []arrow.Record, columnName string) {
 	sch := table.ToArrowSchema()
 	if !sch.HasField(columnName) {