+++ plugin/testing_upsert.go
@@ -3,7 +3,6 @@ package plugin
 import (
 	"context"
 	"fmt"
-	"time"
 
 	"github.com/apache/arrow/go/v13/arrow"
 	"github.com/apache/arrow/go/v13/arrow/array"
@@ -13,7 +12,7 @@ import (
 )
 
 func (s *WriterTestSuite) testUpsertBasic(ctx context.Context) error {
-	tableName := fmt.Sprintf("cq_upsert_basic_%d", time.Now().Unix())
+	tableName := s.tableNameForTest("upsert_basic")
 	table := &schema.Table{
 		Name: tableName,
 		Columns: []schema.Column{
@@ -62,14 +61,15 @@ func (s *WriterTestSuite) testUpsertBasic(ctx context.Context) error {
 	if totalItems != 1 {
 		return fmt.Errorf("expected 1 item, got %d", totalItems)
 	}
-	if diff := RecordDiff(records[0], record); diff != "" {
+	if diff := RecordsDiff(table.ToArrowSchema(), records, []arrow.Record{record}); diff != "" {
 		return fmt.Errorf("record differs: %s", diff)
 	}
 	return nil
 }
 
 func (s *WriterTestSuite) testUpsertAll(ctx context.Context) error {
-	tableName := fmt.Sprintf("cq_upsert_all_%d", time.Now().Unix())
+	const rowsPerRecord = 10
+	tableName := s.tableNameForTest("upsert_all")
 	table := schema.TestTable(tableName, s.genDatOptions)
 	table.Columns = append(table.Columns, schema.Column{Name: "name", Type: arrow.BinaryTypes.String, PrimaryKey: true})
 	if err := s.plugin.writeOne(ctx, &message.WriteMigrateTable{
@@ -79,7 +79,10 @@ func (s *WriterTestSuite) testUpsertAll(ctx context.Context) error {
 	}
 
 	tg := schema.NewTestDataGenerator()
-	normalRecord := tg.Generate(table, schema.GenTestDataOptions{MaxRows: 1})[0]
+	normalRecord := tg.Generate(table, schema.GenTestDataOptions{
+		MaxRows:       rowsPerRecord,
+		TimePrecision: s.genDatOptions.TimePrecision,
+	})
 	if err := s.plugin.writeOne(ctx, &message.WriteInsert{
 		Record: normalRecord,
 	}); err != nil {
@@ -91,16 +94,19 @@ func (s *WriterTestSuite) testUpsertAll(ctx context.Context) error {
 	if err != nil {
 		return fmt.Errorf("failed to readAll: %w", err)
 	}
+	sortRecords(table, records, "id")
+
 	totalItems := TotalRows(records)
-	if totalItems != 1 {
-		return fmt.Errorf("expected 1 item, got %d", totalItems)
+	if totalItems != rowsPerRecord {
+		return fmt.Errorf("expected items after initial insert: %d, got %d", rowsPerRecord, totalItems)
 	}
 
-	if diff := RecordDiff(records[0], normalRecord); diff != "" {
+	if diff := RecordsDiff(table.ToArrowSchema(), records, []arrow.Record{normalRecord}); diff != "" {
 		return fmt.Errorf("record differs after insert: %s", diff)
 	}
 
-	nullRecord := tg.Generate(table, schema.GenTestDataOptions{MaxRows: 1, NullRows: true})[0]
+	tg.Reset()
+	nullRecord := tg.Generate(table, schema.GenTestDataOptions{MaxRows: rowsPerRecord, TimePrecision: s.genDatOptions.TimePrecision, NullRows: true})
 	if err := s.plugin.writeOne(ctx, &message.WriteInsert{
 		Record: nullRecord,
 	}); err != nil {
@@ -112,13 +118,14 @@ func (s *WriterTestSuite) testUpsertAll(ctx context.Context) error {
 	if err != nil {
 		return fmt.Errorf("failed to sync: %w", err)
 	}
+	sortRecords(table, records, "id")
 
 	totalItems = TotalRows(records)
-	if totalItems != 1 {
-		return fmt.Errorf("expected 1 item, got %d", totalItems)
+	if totalItems != rowsPerRecord {
+		return fmt.Errorf("expected items after upsert: %d, got %d", rowsPerRecord, totalItems)
 	}
 
-	if diff := RecordDiff(records[0], nullRecord); diff != "" {
+	if diff := RecordsDiff(table.ToArrowSchema(), records, []arrow.Record{nullRecord}); diff != "" {
 		return fmt.Errorf("record differs after upsert (columns should be null): %s", diff)
 	}