+++ writers/streamingbatchwriter/mocktimer_test.go
@@ -1,17 +1,25 @@
 package streamingbatchwriter
 
 import (
+	"sync"
 	"time"
 
 	"github.com/cloudquery/plugin-sdk/v4/writers"
 )
 
 type mockTicker struct {
-	expire chan time.Time
+	expire  chan time.Time
+	stopped sync.Once
 }
 
 func (t *mockTicker) Stop() {
-	close(t.expire)
+	t.stopped.Do(func() {
+		close(t.expire)
+	})
+}
+
+func (t *mockTicker) Tick() {
+	t.expire <- time.Now()
 }
 
 func (*mockTicker) Reset(time.Duration) {}
@@ -20,12 +28,12 @@ func (t *mockTicker) Chan() <-chan time.Time {
 	return t.expire
 }
 
-func newMockTicker() (writers.TickerFunc, chan<- time.Time) {
+func newMockTicker() (writers.TickerFunc, func()) {
 	expire := make(chan time.Time)
 	t := &mockTicker{
 		expire: expire,
 	}
 	return func(time.Duration) writers.Ticker {
 		return t
-	}, expire
+	}, t.Tick
 }