+++ writers/streamingbatchwriter/streamingbatchwriter.go
@@ -156,13 +156,17 @@ func (w *StreamingBatchWriter) Close(context.Context) error {
 	}
 	w.workersWaitGroup.Wait()
 
-	w.insertWorkers = nil
+	w.insertWorkers = make(map[string]*streamingWorkerManager[*message.WriteInsert])
+	w.migrateWorker = nil
+	w.deleteWorker = nil
+	w.lastMsgType = writers.MsgTypeUnset
 
 	return nil
 }
 
 func (w *StreamingBatchWriter) Write(ctx context.Context, msgs <-chan message.WriteMessage) error {
 	errCh := make(chan error)
+	defer close(errCh)
 
 	go func() {
 		for err := range errCh {
@@ -172,7 +176,7 @@ func (w *StreamingBatchWriter) Write(ctx context.Context, msgs <-chan message.Wr
 
 	for msg := range msgs {
 		msgType := writers.MsgID(msg)
-		if w.lastMsgType != msgType {
+		if w.lastMsgType != writers.MsgTypeUnset && w.lastMsgType != msgType {
 			if err := w.Flush(ctx); err != nil {
 				return err
 			}
@@ -183,12 +187,7 @@ func (w *StreamingBatchWriter) Write(ctx context.Context, msgs <-chan message.Wr
 		}
 	}
 
-	if err := w.Flush(ctx); err != nil {
-		return err
-	}
-
-	close(errCh)
-	return nil
+	return w.Close(ctx)
 }
 
 func (w *StreamingBatchWriter) startWorker(ctx context.Context, errCh chan<- error, msg message.WriteMessage) error {
@@ -248,6 +247,7 @@ func (w *StreamingBatchWriter) startWorker(ctx context.Context, errCh chan<- err
 
 			batchSizeRows: w.batchSizeRows,
 			batchTimeout:  w.batchTimeout,
+			tickerFn:      w.tickerFn,
 		}
 
 		w.workersWaitGroup.Add(1)